#5 [feat] update signUpView
