+++ DropDrug/Resources/Assets.xcassets/OB2.imageset/Contents.json
@@ -1,7 +1,7 @@
 {
   "images" : [
     {
-      "filename" : "placeholder2.jpg",
+      "filename" : "OB2.jpg",
       "idiom" : "universal",
       "scale" : "1x"
     },
