+++ DropDrug/Resources/Assets.xcassets/OB3.imageset/Contents.json
@@ -1,7 +1,7 @@
 {
   "images" : [
     {
-      "filename" : "placeholder1.jpg",
+      "filename" : "OB3.jpg",
       "idiom" : "universal",
       "scale" : "1x"
     },
