+++ DropDrug/Resources/Assets.xcassets/OB4.imageset/Contents.json
@@ -1,7 +1,7 @@
 {
   "images" : [
     {
-      "filename" : "placeholder3.jpg",
+      "filename" : "OB4.jpg",
       "idiom" : "universal",
       "scale" : "1x"
     },
