+++ DropDrug/Sources/SceneDelegate.swift
@@ -17,7 +17,8 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         window = UIWindow(frame: windowScene.coordinateSpace.bounds)
         window?.windowScene = windowScene
  
-        let viewController = TestVC()
+        let viewController = SignUpVC()
+//        let viewController = OnboardingVC()
         
         window?.rootViewController = viewController
         window?.makeKeyAndVisible()
