+++ DropDrug/Sources/ViewControllers/Register/OnboardingVC.swift
@@ -9,10 +9,10 @@ let superViewWidth = UIScreen.main.bounds.width
 class OnboardingVC: UIViewController, UICollectionViewDelegate, UICollectionViewDataSource, UICollectionViewDelegateFlowLayout {
     
     let pages: [OnboardingPage] = [
-        OnboardingPage(title: "마약성분 의약품이 악용돼요", description: "마약성분을 포함한 의약품이 무분별하게 폐기될 경우, 마약류 약물이 유통되거나 악용될 수 있어요.", imageName: "placeholder1"),
-        OnboardingPage(title: "다제내성균이 확산돼요", description: "다제내성균은 여러 종류의 항생제에 내성이 있는 세균을 말해요. 치료할 수 있는 항생제가 적어 감염질환 치료가 어려운데, 잘못된 의약품 폐기로 다제내성균 출현이 유도될 수 있어요.", imageName: "placeholder2"),
-        OnboardingPage(title: "수질 오염이 심각해요", description: "의약품을 제대로 폐기하지 않고 땅에 매립하거나 하수구에 버리게 되면 항생물질 등의 약품 성분이 토양과 지하수, 하천에 유입되면서 수질을 오염시켜요.", imageName: "placeholder3"),
-        OnboardingPage(title: "수질 오염이 심각해요", description: "의약품을 제대로 폐기하지 않고 땅에 매립하거나 하수구에 버리게 되면 항생물질 등의 약품 성분이 토양과 지하수, 하천에 유입되면서 수질을 오염시켜요.", imageName: "placeholder3")
+        OnboardingPage(title: "사용기한이 지난 약,\n어디에 버리시나요?", description: "2018년 건강보험심사평가원 조사에 따르면, 응답자의 절반 이상이 폐의약품을 쓰레기통, 하수구, 변기 등에 버리고 있으며, 약국, 보건소, 주민센터의 수거함에 배출한다는 응답은 8%에 불과했습니다. 폐의약품 처리 방법을 모른다는 응답자는 무려 74.1%에 달했습니다.", imageName: "OB1"),
+        OnboardingPage(title: "쓰레기통이나 변기에\n버리면 왜 안 되나요?", description: "국립환경과학원의 조사에 따르면, 우리나라 하천에서는 소염진통제, 항생제 성분이 검출되고 있으며, 이로 인해 기형 물고기와 항생제 내성을 가진 슈퍼 박테리아가 발견됐습니다. 무심코 버린 폐의약품이 수질과 생태계를 오염시키고 있는 것입니다.", imageName: "OB2"),
+        OnboardingPage(title: "폐의약품은 이렇게 버리세요", description: "알약: 약만 따로 모아 배출합니다.\n가루약: 포장지 그대로 배출합니다.\n물약 및 시럽: 밀봉한 용기에 모아 배출합니다.\n연고, 안약, 흡입제 등: 특수 용기는 그대로 배출합니다.", imageName: "OB3"),
+        OnboardingPage(title: "어디에 버려야 하나요?", description: "폐의약품은 보건소나 주민센터의 수거함에 직접 배출할 수 있으며, 2024년부터는 우체통을 통해서도 손쉽게 배출할 수 있습니다. 가까운 배출 장소는 드롭드락 앱에서 확인해 보세요!", imageName: "OB4")
     ]
     
     private var collectionView: UICollectionView!
@@ -143,21 +143,22 @@ class OnboardingCollectionViewCell: UICollectionViewCell {
         }
         
         titleLabel.textColor = UIColor(named: "pink")
-        titleLabel.font = UIFont.boldSystemFont(ofSize: 22)
+        titleLabel.font = UIFont.ptdSemiBoldFont(ofSize: 22)
         titleLabel.textAlignment = .center
+        titleLabel.numberOfLines = 0
         textOverlayView.addSubview(titleLabel)
         titleLabel.snp.makeConstraints { make in
-            make.top.equalTo(textOverlayView.snp.top).offset(superViewHeight * 0.05)
+            make.top.equalTo(textOverlayView.snp.top).offset(superViewHeight * 0.04)
             make.leading.trailing.equalToSuperview()
         }
         
         descriptionLabel.textColor = .white
-        descriptionLabel.font = UIFont.systemFont(ofSize: 16)
+        descriptionLabel.font = UIFont.ptdRegularFont(ofSize: 16)
         descriptionLabel.textAlignment = .center
         descriptionLabel.numberOfLines = 0
         textOverlayView.addSubview(descriptionLabel)
         descriptionLabel.snp.makeConstraints { make in
-            make.top.equalTo(titleLabel.snp.bottom).offset(superViewHeight * 0.05)
+            make.top.equalTo(titleLabel.snp.bottom).offset(superViewHeight * 0.04)
             make.leading.trailing.equalToSuperview().inset(superViewWidth * 0.1)
         }
     }
