+++ DropDrug/Sources/ViewControllers/Register/SplashVC.swift
@@ -10,7 +10,7 @@ class SplashVC : UIViewController {
     private lazy var titleLabel: UILabel = {
         let label = UILabel()
         label.text = "Drop Drug"
-        label.font = UIFont.systemFont(ofSize: 50)
+        label.font = UIFont.roRegularFont(ofSize: 50)
         label.textColor = .white
         label.textAlignment = .center
         return label
@@ -20,7 +20,7 @@ class SplashVC : UIViewController {
         super.viewDidLoad()
         setupViews()
         setConstraints()
-        DispatchQueue.main.asyncAfter(deadline: .now() + 2.0) {self.navigateToOnBoaringScreen()}
+        DispatchQueue.main.asyncAfter(deadline: .now() + 2.0) {self.navigateToOnBoaringScreen()} //일단 분기 처리 없이 온보딩 페이지로 넘어가도록
 //        DispatchQueue.main.asyncAfter(deadline: .now() + 2.0) {
 //            if self.getRefreshToken() != nil {
 //                        // refreshToken 이 있으면 메인 화면으로 이동
@@ -33,8 +33,9 @@ class SplashVC : UIViewController {
 //    }
     
     func setupViews() {
-        view.addSubview(titleLabel)
         view.backgroundColor = UIColor(named: "skyblue")
+        view.addSubview(titleLabel)
+        
     }
     
     func navigateToMainScreen() {
