+++ DropDrug/Sources/ViewControllers/Register/loginVC.swift
@@ -1,0 +1,8 @@
+// Copyright © 2024 RT4. All rights reserved
+
+import UIKit
+import SnapKit
+
+class loginVC : UIViewController {
+    
+}
